+++ tests/testthat/test-create_dt.R
@@ -1,0 +1,5 @@
+test_that("create_dt works", {
+
+  dt <- create_dt(dat = mtcars)
+  testthat::expect_true(methods::is(dt,"datatables"))
+})